Return a partial ODP client object if not found in Hydra
This prevents the admin UI client list from breaking if you've imported, for example, an ODP production database to local, when your local Hydra instance does not have all the production clients.

This does not claim to deal with what might happen if you try to edit such an object.